Fix: Miscellaneous fixes

+++ htdocs/install/mysql/migration/3.4.0-3.5.0.sql
@@ -359,5 +359,5 @@ ALTER TABLE llx_societe ADD skype VARCHAR(255) AFTER email;
 ALTER TABLE llx_adherent ADD skype VARCHAR(255) AFTER email;
 
 -- multi-rib
-ALTER TABLE llx_societe_rib ADD default_rib TINYINT NOT NULL DEFAULT 0 AFTER owner_address;
+ALTER TABLE llx_societe_rib ADD default_rib smallint NOT NULL DEFAULT 0 AFTER owner_address;
 UPDATE llx_societe_rib SET default_rib = 1;
